+++ client/index.jsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import $ from 'jquery';
 import Gallery from './src/Gallery.jsx';
 import Description from './src/Description.jsx';
+import Fit from './src/Fit.jsx';
 import Sizes from './src/Sizes.jsx';
 import Colors from './src/Colors.jsx';
 import AddToBag from './src/AddToBag.jsx';
